pyproject.toml
[project]
name = "docfetcher"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyyaml", "python-frontmatter"]

[tool.setuptools]
package-dir = {"" = "src"}